Autplay Enabled - Jump to a specific slide fix
Before fix:
1. Enable Autoplay
2. Disable Loop
3. Go to last slide and wait for "end" state.
4. Execute $elmt.vegas('jump', any_slide_goes_here);
Result: Vegas will jump to specific slide but ***will not play*** next slide. It appears like autoplay is killed (which is sort of right due to _end function).

After fix:
1. Enable Autoplay
2. Disable Loop
3. Go to last slide and wait for "end" state.
4. Execute $elmt.vegas('jump', any_slide_goes_here);
Result: Vegas will jump to specific slide and after that it will autoplay next slide as it supposed to do.